\name{achievement}
\alias{achievement}
\docType{data}
\title{School achievement data}
\description{
Achievement data for a group of Austrian school children
}
\usage{
achievement
}
\format{
  A data frame with 109 observations on the following 7 variables.
  \describe{
  \item{Gen}{gender of child where 0 is male and 1 is female}
  \item{Age}{age in months}
  \item{IQ}{iq score}
  \item{math1}{test score on mathematics computation}
  \item{math2}{test score on mathematics problem solving}
  \item{read1}{test score on reading speed}
  \item{read2}{test score on reading comprehension}
  }
}
\source{
Abraham, B., and Ledolter, J. (2006), Introduction to Regression Modeling, 
Duxbury.
}

\keyword{datasets}
