\name{baseball.1964}
\alias{baseball.1964}
\docType{data}
\title{Team records in the 1964 National League baseball season}
\description{
Head to head records for all teams in the 1964 National League baseball
season.  Teams are coded as Cincinnati (1), Chicago (2), Houston (3),
Los Angeles (4), Milwaukee (5), New York (6), Philadelphia (7),
Pittsburgh (8), San Francisco (9), and St. Louis (10).
}
\usage{
baseball.1964
}
\format{
   A data frame with 45 observations on the following 4 variables.
  \describe{
  \item{Team.1}{Number of team 1}
  \item{Team.2}{Number of team 2}
  \item{Wins.Team1}{Number of games won by team 1}
  \item{Wins.Team2}{Number of games won by team 2}
  }
}
\source{
www.baseball-reference.com website.
}

\keyword{datasets}
