\name{birdextinct}
\alias{birdextinct}
\docType{data}
\title{Bird measurements from British islands}
\description{
Measurements on breedings pairs of landbird species were collected from 16
islands about Britain over several decades.
}
\usage{
birdextinct
}
\format{
A data frame with 62 observations on the following 5 variables.
  \describe{
  \item{species}{name of bird species}
  \item{time}{average time of extinction on the islands}
  \item{nesting}{average number of nesting pairs}
  \item{size}{size of the species, 1 or 0 if large or small}
  \item{status}{staus of the species, 1 or 0 if resident or migrant}
  }
}
\source{
Pimm, S., Jones, H., and Diamond, J. (1988),
On the risk of extinction,
American Naturalists, 132, 757-785.
}

\keyword{datasets}
