\name{calculus.grades}
\alias{calculus.grades}
\docType{data}
\title{Calculus grades dataset}
\description{
Grades and other variables collected for a 
sample of calculus students.
}
\usage{
calculus.grades
}
\format{
   A data frame with 100 observations on the following 3 variables.
  \describe{
  \item{grade}{indicates if student received a A or B in class}
  \item{prev.grade}{indicates if student received a A in prerequisite math class}
  \item{act}{score on the ACT math test}
  }
}
\source{Collected by a colleague of the author at his university.}
\keyword{datasets}
