\name{donner}
\alias{donner}
\docType{data}
\title{Donner survival study}
\description{
Data contains the age, gender and survival status for 45 members of the Donner Party
who experienced difficulties in crossing the Sierra Nevada mountains in California.
}
\usage{
donner
}
\format{
A data frame with 45 observations on the following 3 variables.
  \describe{
  \item{age}{age of person}
  \item{male}{gender that is 1 (0) if person is male (female)}
  \item{survival}{survival status, 1 or 0 if person survived or died}
  }
}
\source{Grayson, D. (1960), Donner party deaths:  a demographic
assessment, Journal of Anthropological Assessment, 46, 223-242.}

\keyword{datasets}
