\name{election.2008}
\alias{election.2008}
\docType{data}
\title{Poll data from 2008 U.S. Presidential Election}
\description{
Results of recent state polls in the 2008 United States Presidential
Election between Barack Obama and John McCain.
}
\usage{
election.2008
}
\format{
  A data frame with 51 observations on the following 4 variables.
  \describe{
  \item{State}{name of the state}
  \item{M.pct}{percentage of poll survey for McCain}
  \item{O.pct}{precentage of poll survey for Obama}
  \item{EV}{number of electoral votes}
  }
}
\source{Data collected by author in November 2008 from www.cnn.com website.}
\keyword{datasets}
