\name{footballscores}
\alias{footballscores}
\docType{data}
\title{Game outcomes and point spreads for American football}
\description{
Game outcomes and point spreads for 672 professional
American football games.
}
\usage{
footballscores
}
\format{
   A data frame with 672 observations on the following 8 variables.
  \describe{
  \item{year}{year of game}
  \item{home}{indicates if favorite is the home team}
  \item{favorite}{score of favorite team}
  \item{underdog}{score of underdog team}
  \item{spread}{point spread}
  \item{favorite.name}{name of favorite team}
  \item{underdog.name}{name of underdog team}
  \item{week}{week number of the season}
  }
}
\source{Gelman, A., Carlin, J., Stern, H., and Rubin, D. (2003),
Bayesian Data Analysis, Chapman and Hall.}
\keyword{datasets}
