\name{iowagpa}
\alias{iowagpa}
\docType{data}
\title{Admissions data for an university}
\description{
Students at a major university are categorized with respect to their
high school rank and their ACT score.  For each combination of high school rank and
ACT score, one records the mean grade point average (GPA).
}
\usage{
iowagpa
}
\format{
 A data frame with 40 observations on the following 4 variables.
  \describe{
  \item{gpa}{mean grade point average}
  \item{n}{sample size}
  \item{HSR}{high school rank}
  \item{ACT}{act score}
  }
}
\source{Albert, J. (1994), A Bayesian approach to estimation of GPA's
of University of Iowa freshmen under order restrictions,
Journal of Educational Statistics, 19, 1-22.}
\keyword{datasets}
