\name{marathontimes}
\alias{marathontimes}
\docType{data}
\title{Marathon running times}
\description{
Running times in minutes for twenty male runners between the ages 20 and 29
who ran the New York Marathon.
}
\usage{
marathontimes
}
\format{
A data frame with 20 observations on the following 1 variable.
 \describe{
  \item{time}{running time}
  }
}
\source{www.nycmarathon.org website.}
\keyword{datasets}
