\name{sluggerdata}
\alias{sluggerdata}
\docType{data}
\title{Hitting statistics for ten great baseball players}
\description{
Career hitting statistics for ten great baseball players
}

\usage{
sluggerdata
}
\format{
 A data frame with 199 observations on the following 13 variables.
  \describe{
  \item{Player}{names of the ballplayer}
  \item{Year}{season played}
  \item{Age}{age of the player during the season}
  \item{G}{games played}
  \item{AB}{number of at-bats}
  \item{R}{number of runs scored}
  \item{H}{number of hits}
  \item{X2B}{number of doubles}
  \item{X3B}{number of triples}
  \item{HR}{number of home runs}
  \item{RBI}{runs batted in}
  \item{BB}{number of base on balls}
  \item{SO}{number of strikeouts}
  }
}
\source{Sean Lahman's baseball database from www.baseball1.com.}
\keyword{datasets}
