\name{soccergoals}
\alias{soccergoals}
\docType{data}
\title{Goals scored by professional soccer team}
\description{
Number of goals scored by a single professional soccer team
during the 2006 Major League Soccer season}

\usage{
soccergoals
}
\format{
  A data frame with 35 observations on the following 1 variable.
  \describe{
  \item{goals}{number of goals scored}
  }
}
\source{Collected by author from the www.espn.com website.}
\keyword{datasets}
