\name{bermuda.grass}
\alias{bermuda.grass}
\docType{data}
\title{Bermuda grass experiment data}
\description{
Yields of bermuda grass for a factorial design of nutrients
nitrogen,
phosphorus, and potassium.
}
\usage{
bermuda.grass
}
\format{
   A data frame with 64 observations on the following 4 variables.
  \describe{
  \item{y}{yield of bermuda grass in tons per acre}
  \item{Nit}{level of nitrogen}
  \item{Phos}{level of phosphorus}
  \item{Pot}{level of potassium}
  }
}
\source{
McCullagh, P., and Nelder, J. (1989), Generalized Linear Models,
Chapman and Hall.
}
\keyword{datasets}
