\name{bermuda.grass}
\alias{bermuda.grass}
\title{Bermuda grass experiment data}
\description{
Yields of bermuda grass for a factorial design of nutrients
nitrogen,
phosphorus, and potassium.
}
\usage{
bermuda.grass
}
\format{
  \item{y}{yield of bermuda grass in tons per acre}
  \item{Nit}{level of nitrogen}
  \item{Phos}{level of phosphorus}
  \item{Pot}{level of potassium}
}

\keyword{datasets}
