\name{birdextinct}
\alias{birdextinct}
\title{Bird measurements from British islands}
\description{
Measurements on breedings pairs of landbird species were collected from 16
islands about Britain over several decades.
}
\usage{
birdextinct
}
\format{
  \item{species}{name of bird species}
  \item{time}{average time of extinction on the islands}
  \item{nesting}{average number of nesting pairs}
  \item{size}{size of the species, 1 or 0 if large or small}
  \item{status}{staus of the species, 1 or 0 if resident or migrant}
}

\keyword{datasets}
