\name{birthweight}
\alias{birthweight}
\docType{data}
\title{Birthweight regression study}
\description{
Dobson describes a study where one is interested in predicting 
a baby's birthweight based on the gestational age and the baby's
gender.
}
\usage{
birthweight
}
\format{
  A data frame with 24 observations on the following 3 variables.
  \describe{
  \item{age}{gestational age in weeks}
  \item{gender}{gender of the baby where 0 (1) is male (female)}
  \item{weight}{birthweight of baby in grams}
  }
}
\source{Dobson, A. (2001), An Introduction to Generalized Linear Models, New York:
Chapman and Hall.}

\keyword{datasets}
