\name{birthweight}
\alias{birthweight}
\title{Birthweight regression study}
\description{
Dobson describes a study where one is interested in predicting 
a baby's birthweight based on the gestational age and the baby's
gender.
}
\usage{
birthweight
}
\format{
  \item{age}{gestational age in weeks}
  \item{gender}{gender of the baby where 0 (1) is male (female)}
  \item{weight}{birthweight of baby in grams}
}

\keyword{datasets}
