\name{breastcancer}
\alias{breastcancer}
\docType{data}
\title{Survival experience of women with breast cancer under treatment}
\description{
Collett (1994) describes a study to evaluate the effectiveness of a
histochemical marker in predicting the survival experience of women with
breast cancer.
}
\usage{
breastcancer
}
\format{
  A data frame with 45 observations on the following 3 variables.
  \describe{
  \item{time}{survival time in months}
  \item{status}{censoring indicator where 1 (0) indicates a complete (censored) survival time}
  \item{stain}{indicates by a 0 (1) if tumor was negatively (positively) stained}
  }
}
\source{Collett, D. (1994), Modelling Survival Data in Medical Research, London:
Chapman and Hall.}

\keyword{datasets}
