\name{breastcancer}
\alias{breastcancer}
\title{Survival experience of women with breast cancer under treatment}
\description{
Collett (1994) describes a study to evaluate the effectiveness of a
histochemical marker in predicting the survival experience of women with
breast cancer.
}
\usage{
breastcancer
}
\format{
  \item{time}{survival time in months}
  \item{status}{censoring indicator where 1 (0) indicates a complete (censored) survival time}
  \item{stain}{indicates by a 0 (1) if tumor was negatively (positively) stained}
}

\keyword{datasets}
