\name{cancermortality}
\alias{cancermortality}
\docType{data}
\title{Cancer mortality data}
\description{
Number of cancer deaths and number at risk for 20 cities in Missouri.
}
\usage{
cancermortality
}
\format{
A data frame with 20 observations on the following 2 variables.
  \describe{
  \item{y}{number of cancer deaths}
  \item{n}{number at risk}
  }
}
\source{Tsutakawa, R., Shoop, G., and Marienfeld, C. (1985),
Empirical Bayes Estimation
of Cancer Mortality Rates, Statistics in Medicine, 4, 201-212.
}

\keyword{datasets}
