\name{chemotherapy}
\alias{chemotherapy}
\title{Chemotherapy treatment effects on ovarian cancer}
\description{
Edmunson et al (1979) studied the effect of different chemotherapy
treatments following surgical treatment of ovarian cancer.
}
\usage{
chemotherapy
}
\format{
  \item{patient}{patient number}
  \item{time}{survival time in days following treatment}
  \item{status}{indicates if time is censored (0) or actually observed (1)}
  \item{treat}{control group (0) or treatment group (1)}
  \item{age}{age of the patient}
}

\keyword{datasets}
