\name{darwin}
\alias{darwin}
\docType{data}
\title{Darwin's data on plants}
\description{
Fifteen differences of the heights of cross and self fertilized plants quoted by
Fisher (1960)}

\usage{
darwin
}
\format{
A data frame with 15 observations on the following 1 variable.
  \describe{
  \item{difference}{difference of heights of two types of plants}
  }
}
\source{Fisher, R. (1960), Statistical Methods for Research Workers,
Edinburgh: Oliver and Boyd.}
\keyword{datasets}
