\name{donner}
\alias{donner}
\title{Donner survival study}
\description{
Data contains the age, gender and survival status for 45 members of the Donner Party
who experienced difficulties in crossing the Sierra Nevada mountains in California.
}
\usage{
donner
}
\format{
  \item{age}{age of person}
  \item{male}{gender that is 1 (0) if person is male (female)}
  \item{survival}{survival status, 1 or 0 if person survived or died}
}

\keyword{datasets}
