\name{election}
\alias{election}
\title{Florida election data}
\description{
For each of the Florida counties in the 2000 presidential election, the number of
votes for George Bush, Al Gore, and Pat Buchanan is recorded.  Also the number of
votes for the minority candidate Ross Perot in the 1992 presidential election is
recorded.
}
\usage{
election
}
\format{
  \item{county}{name of Florida county}
  \item{perot}{number of votes for Ross Perot in 1992 election}
  \item{gore}{number of votes for Al Gore in 2000 election}
  \item{bush}{number of votes for George Bush in 2000 election}
  \item{buchanan}{number of votes for Pat Buchanan in 2000 election}
}

\keyword{datasets}
