\name{footballscores}
\alias{footballscores}
\title{Game outcomes and point spreads for American football}
\description{
Game outcomes and point spreads for 672 professional
American football games.
}
\usage{
footballscores
}
\format{
  \item{year}{year of game}
  \item{home}{indicates if favorite is the home team}
  \item{favorite}{score of favorite team}
  \item{underdog}{score of underdog team}
  \item{spread}{point spread}
  \item{favorite.name}{name of favorite team}
  \item{underdog.name}{name of underdog team}
  \item{week}{week number of the season}
}
}

\keyword{datasets}
