\name{hearttransplants}
\alias{hearttransplants}
\docType{data}
\title{Heart transplant mortality data}
\description{
The number of deaths within 30 days of heart transplant surgery for 94 U.S. hospitals
that performed at least 10 heart transplant surgeries.  Also the exposure, the expected number of deaths, is 
recorded for each hospital.}

\usage{
hearttransplants
}
\format{
A data frame with 94 observations on the following 2 variables.
  \describe{
  \item{e}{expected number of deaths (the exposure)}
  \item{y}{observed number of deaths within 30 days of heart transplant surgery}
  }
}
\source{Christiansen, C. and Morris, C. (1995), Fitting and checking
a two-level Poisson model:  modeling patient mortality rates in heart
transplant patients, in Berry, D. and Stangl, D., eds, Bayesian
Biostatistics, Marcel Dekker.}

\keyword{datasets}
