\name{hearttransplants}
\alias{hearttransplants}
\title{Heart transplant mortality data}
\description{
The number of deaths within 30 days of heart transplant surgery for 84 U.S. hospitals
that performed at least 10 heart transplant surgeries.  Also the exposure, the expected number of deaths, is 
recorded for each hospital.}

\usage{
hearttransplants
}
\format{
  \item{e}{expected number of deaths (the exposure)}
  \item{y}{observed number of deaths within 30 days of heart transplant surgery}
}

\keyword{datasets}
