\name{iowagpa}
\alias{iowagpa}
\title{Admissions data for an university}
\description{
Students at a major university are categorized with respect to their
high school rank and their ACT score.  For each combination of high school rank and
ACT score, one records the mean grade point average (GPA).
}
\usage{
iowagpa
}
\format{
  \item{gpa}{mean grade point average}
  \item{n}{sample size}
  \item{HSR}{high school rank}
  \item{ACT}{act score}
}

\keyword{datasets}
