\name{jeter2004}
\alias{jeter2004}
\title{Hitting data for Derek Jeter}
\description{
Batting data for the baseball player Derek Jeter for all 154 games in the 2004 season.}

\usage{
jeter2004
}
\format{
  \item{Game}{the game number}
  \item{AB}{the number of at-bats}
  \item{R}{the number of runs scored}
  \item{H}{the number of hits}
  \item{X2B}{the number of doubles}
  \item{X3B}{the number of triples}
  \item{HR}{the number of home runs}
  \item{RBI}{the number of runs batted in}
  \item{BB}{the number of walks}
  \item{SO}{the number of strikeouts}
}

\keyword{datasets}
