\name{puffin}
\alias{puffin}
\docType{data}
\title{Bird measurements from British islands}
\description{
Measurements on breedings of the common puffin on different
habits at Great Island, Newfoundland.
}
\usage{
puffin
}
\format{
 A data frame with 38 observations on the following 5 variables.
  \describe{
  \item{Nest}{nesting frequency (burrows per 9 square meters)}
  \item{Grass}{grass cover (percentage)}
  \item{Soil}{mean soil depth (in centimeters)}
  \item{Angle}{angle of slope (in degrees)}
  \item{Distance}{distance from cliff edge (in meters)}
  }
}
\source{Peck, R., Devore, J., and Olsen, C. (2005), Introduction to Statistics 
And Data Analysis, Thomson Learning.}
\keyword{datasets}
