\name{schmidt}
\alias{schmidt}
\docType{data}
\title{Batting data for Mike Schmidt}
\description{
Batting statistics for the baseball player Mike Schmidt
during all the seasons of his career.
}
\usage{
schmidt
}
\format{
  A data frame with 18 observations on the following 14 variables.
  \describe{
  \item{Year}{year of the season}
  \item{Age}{Schmidt's age that season}
  \item{G}{games played}
  \item{AB}{at-bats}
  \item{R}{runs scored}
  \item{H}{number of hits}
  \item{X2B}{number of doubles}
  \item{X3B}{number of triples}
  \item{HR}{number of home runs}
  \item{RBI}{number of runs batted in}
  \item{SB}{number of stolen bases}
  \item{CS}{number of times caught stealing}
  \item{BB}{number of walks}
  \item{SO}{number of strikeouts}
  }
}
\source{Sean Lahman's baseball database from www.baseball1.com.}
\keyword{datasets}
