\name{sluggerdata}
\alias{sluggerdata}
\title{Hitting statistics for ten great baseball players}
\description{
Career hitting statistics for ten great baseball players
}

\usage{
sluggerdata
}
\format{
  \item{Player}{names of the ballplayer}
  \item{Year}{season played}
  \item{Age}{age of the player during the season}
  \item{G}{games played}
  \item{AB}{number of at-bats}
  \item{R}{number of runs scored}
  \item{H}{number of hits}
  \item{X2B}{number of doubles}
  \item{X3B}{number of triples}
  \item{HR}{number of home runs}
  \item{RBI}{runs batted in}
  \item{BB}{number of base on balls}
  \item{SO}{number of strikeouts}
}

\keyword{datasets}
