\name{soccergoals}
\alias{soccergoals}
\title{Goals scored by professional soccer team}
\description{
Number of goals scored by a single professional soccer team
during the 2006 Major League Soccer season}

\usage{
soccergoals
}
\format{
  \item{goals}{number of goals scored}
}

\keyword{datasets}
