\name{stanfordheart}
\alias{stanfordheart}
\docType{data}
\title{Data from Stanford Heart Transplanation Program}
\description{
Heart transplant data for 82 patients from Stanford Heart Transplanation Program}

\usage{
stanfordheart
}
\format{
A data frame with 82 observations on the following 4 variables.
  \describe{
  \item{survtime}{survival time in months}
  \item{transplant}{variable that is 1 or 0 if patient had transplant or not}
  \item{timetotransplant}{time a transplant patient waits for operation}
  \item{state}{variable that is 1 or 0 if time is censored or not}
  }
}
\source{Turnbull, B., Brown, B. and Hu, M. (1974), Survivorship 
analysis of heart transplant data, Journal of the 
 American Statistical Association, 69, 74-80.}
\keyword{datasets}
