\name{stanfordheart}
\alias{stanfordheart}
\title{Data from Stanford Heart Transplanation Program}
\description{
Heart transplant data for 82 patients from Stanford Heart Transplanation Program}

\usage{
stanfordheart
}
\format{
  \item{survtime}{survival time in months}
  \item{transplant}{variable that is 1 or 0 if patient had transplant or not}
  \item{timetotransplant}{time a transplant patient waits for operation}
  \item{state}{variable that is 1 or 0 if time is censored or not}
}

\keyword{datasets}
