\name{strikeout}
\alias{strikeout}
\docType{data}
\title{Baseball strikeout data}
\description{
For all professional baseball players in the 2004 season, dataset gives the
number of strikeouts and at-bats when runners are in scoring position and when
runners are not in scoring position.
}
\usage{
strikeout
}
\format{
  A data frame with 438 observations on the following 4 variables.
  \describe{
  \item{r}{number of strikeouts of player when runners are not in scoring position}
  \item{n}{number of at-bats of player when runners are not in scoring position}
  \item{s}{number of strikeouts of player when runners are in scoring position}
  \item{m}{number of at-bats of player when runners are in scoring position}
  }
}
\source{Collected from www.espn.com website.}
\keyword{datasets}
