\name{studentdata}
\alias{studentdata}
\docType{data}
\title{Student dataset}
\description{
Answers to a sheet of questions given to a large number
of students in introductory statistics classes
}
\usage{
studentdata
}
\format{
 A data frame with 657 observations on the following 11 variables.
  \describe{
  \item{Student}{student number}
  \item{Height}{height in inches}
  \item{Gender}{gender}
  \item{Shoes}{number of pairs of shoes owned}
  \item{Number}{number chosen between 1 and 10}
  \item{Dvds}{name of movie dvds owned}
  \item{ToSleep}{time the person went to sleep the previous night (hours past midnight)}
  \item{WakeUp}{time the person woke up the next morning}
  \item{Haircut}{cost of last haircut including tip}
  \item{Job}{number of hours working on a job per week}
  \item{Drink}{usual drink at suppertime among milk, water, and pop}
  }
}

\source{Collected by the author during the Fall 2006 semester.}
\keyword{datasets}
